+++ NEWS.md
@@ -5,6 +5,7 @@
 	* Utility function `quiet()` to run any functions without warning/error/print messages
 	* `cohort_creation()` to generate cohort and show number (%) of entries at each inclusion/exclusion step
   * `compare_sets()` to find the number of unique and common elements in two sets
+  * `create_ntiles()` to bin numeric variables into user-specified quantiles
 
 * **Bug fixes:**
   * Minor bug fix in `plot_over_time` to allow custom line color without specifying color grouping variable